adding boolean to turn sound off
workaround for game not starting if flash doesnt have adequate permissions.

added a var to turn sound off programmatically to debug the flash sound manager which wont run with out express permission if on a local file system (which is the case for chrome extensions it appears)

This var can be turned in to a button which the user can turn on and off, and can be greyed out if the flash permisions are unavailible (ie, the user needs to visit http://www.macromedia.com/support/documentation/en/flashplayer/help/settings_manager04.html and turn on sound.

Ideally the work around should be to bring the user to enable sound for this webapp.